fix: change product ID type from Long to Integer in ICartService methods

+++ backend/src/main/java/com/farma_ya/service/ICartService.java
@@ -11,9 +11,9 @@ public interface ICartService {
 
     Cart getCartByUser(User user);
 
-    Cart addToCart(User user, Long productId, int quantity);
+    Cart addToCart(User user, Integer productId, int quantity);
 
-    Cart removeFromCart(User user, Long productId);
+    Cart removeFromCart(User user, Integer productId);
 
     Cart clearCart(User user);
 }
